+++ bin-proto/src/types/mod.rs
@@ -4,6 +4,7 @@ mod array;
 mod collections;
 mod cstring;
 mod marker;
+mod net;
 mod numerics;
 mod option;
 mod smart_ptr;