+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # v0.3.0
 - Implement `ExternallyLengthPrefixed` on `HashMap` and `BTreeMap` instead of implementing `Protocol`
 - Do not implement `Protocol` on `char`, `range`
+- Implement protocol on `Ipv4Addr` and `Ipv6Addr`, `(T0, )`, `()`, `Box`
 # v0.2.0
 - Add context to all parse functions
 - Remove `#[repr(...)]`, instead specify repr in `#[protocol(discriminant = "...")]`